poc/requirements.txt
anyio==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.7
httpx==0.22.0
hyperframe==6.0.1
idna==3.3
kafka-python==2.0.2
pytchat==0.5.5
rfc3986==1.5.0
sniffio==1.2.0
